requirements.txt
fastapi
uvicorn[standard]
streamlit>=1.24.0
python-dotenv
pydantic
requests
pandas>=2.0.3
numpy>=1.26.0
matplotlib>=3.7.2
seaborn>=0.12.2
scikit-learn>=1.3.0
pytest>=7.4.0